microservices/users/requirements.txt
grpcio
grpcio-tools
protobuf
python-dotenv